bthrm - Move code into lib to boot faster if disabled
