+++ apps/bthrm/metadata.json
@@ -15,6 +15,7 @@
     {"name":"bthrm.0.boot.js","url":"boot.js"},
     {"name":"bthrm.img","url":"app-icon.js","evaluate":true},
     {"name":"bthrm.settings.js","url":"settings.js"},
+    {"name":"bthrm","url":"lib.js"},
     {"name":"bthrm.default.json","url":"default.json"}
   ]
 }
